+++ google/genai/caches.py
@@ -116,7 +116,11 @@ def _CreateCachedContentConfig_to_mldev(
     )
 
   if getv(from_object, ['tool_config']) is not None:
-    setv(parent_object, ['toolConfig'], getv(from_object, ['tool_config']))
+    setv(
+        parent_object,
+        ['toolConfig'],
+        _ToolConfig_to_mldev(getv(from_object, ['tool_config']), to_object),
+    )
 
   if getv(from_object, ['kms_key_name']) is not None:
     raise ValueError('kms_key_name parameter is not supported in Gemini API.')
@@ -289,6 +293,53 @@ def _FileData_to_mldev(
 
   if getv(from_object, ['mime_type']) is not None:
     setv(to_object, ['mimeType'], getv(from_object, ['mime_type']))
+
+  return to_object
+
+
+def _FunctionCall_to_mldev(
+    from_object: Union[dict[str, Any], object],
+    parent_object: Optional[dict[str, Any]] = None,
+) -> dict[str, Any]:
+  to_object: dict[str, Any] = {}
+  if getv(from_object, ['id']) is not None:
+    setv(to_object, ['id'], getv(from_object, ['id']))
+
+  if getv(from_object, ['args']) is not None:
+    setv(to_object, ['args'], getv(from_object, ['args']))
+
+  if getv(from_object, ['name']) is not None:
+    setv(to_object, ['name'], getv(from_object, ['name']))
+
+  if getv(from_object, ['partial_args']) is not None:
+    raise ValueError('partial_args parameter is not supported in Gemini API.')
+
+  if getv(from_object, ['will_continue']) is not None:
+    raise ValueError('will_continue parameter is not supported in Gemini API.')
+
+  return to_object
+
+
+def _FunctionCallingConfig_to_mldev(
+    from_object: Union[dict[str, Any], object],
+    parent_object: Optional[dict[str, Any]] = None,
+) -> dict[str, Any]:
+  to_object: dict[str, Any] = {}
+  if getv(from_object, ['mode']) is not None:
+    setv(to_object, ['mode'], getv(from_object, ['mode']))
+
+  if getv(from_object, ['allowed_function_names']) is not None:
+    setv(
+        to_object,
+        ['allowedFunctionNames'],
+        getv(from_object, ['allowed_function_names']),
+    )
+
+  if getv(from_object, ['stream_function_call_arguments']) is not None:
+    raise ValueError(
+        'stream_function_call_arguments parameter is not supported in Gemini'
+        ' API.'
+    )
 
   return to_object
 
@@ -516,8 +567,10 @@ def _Part_to_mldev(
     parent_object: Optional[dict[str, Any]] = None,
 ) -> dict[str, Any]:
   to_object: dict[str, Any] = {}
-  if getv(from_object, ['function_call']) is not None:
-    setv(to_object, ['functionCall'], getv(from_object, ['function_call']))
+  if getv(from_object, ['media_resolution']) is not None:
+    setv(
+        to_object, ['mediaResolution'], getv(from_object, ['media_resolution'])
+    )
 
   if getv(from_object, ['code_execution_result']) is not None:
     setv(
@@ -534,6 +587,13 @@ def _Part_to_mldev(
         to_object,
         ['fileData'],
         _FileData_to_mldev(getv(from_object, ['file_data']), to_object),
+    )
+
+  if getv(from_object, ['function_call']) is not None:
+    setv(
+        to_object,
+        ['functionCall'],
+        _FunctionCall_to_mldev(getv(from_object, ['function_call']), to_object),
     )
 
   if getv(from_object, ['function_response']) is not None:
@@ -565,6 +625,28 @@ def _Part_to_mldev(
 
   if getv(from_object, ['video_metadata']) is not None:
     setv(to_object, ['videoMetadata'], getv(from_object, ['video_metadata']))
+
+  return to_object
+
+
+def _ToolConfig_to_mldev(
+    from_object: Union[dict[str, Any], object],
+    parent_object: Optional[dict[str, Any]] = None,
+) -> dict[str, Any]:
+  to_object: dict[str, Any] = {}
+  if getv(from_object, ['function_calling_config']) is not None:
+    setv(
+        to_object,
+        ['functionCallingConfig'],
+        _FunctionCallingConfig_to_mldev(
+            getv(from_object, ['function_calling_config']), to_object
+        ),
+    )
+
+  if getv(from_object, ['retrieval_config']) is not None:
+    setv(
+        to_object, ['retrievalConfig'], getv(from_object, ['retrieval_config'])
+    )
 
   return to_object
 
@@ -1047,15 +1129,6 @@ class Caches(_api_module.BaseModule):
   def _list(
       self, *, config: Optional[types.ListCachedContentsConfigOrDict] = None
   ) -> types.ListCachedContentsResponse:
-    """Lists cached content configurations.
-
-    .. code-block:: python
-
-      cached_contents = client.caches.list(config={'page_size': 2})
-      for cached_content in cached_contents:
-        print(cached_content)
-    """
-
     parameter_model = types._ListCachedContentsParameters(
         config=config,
     )
@@ -1114,9 +1187,28 @@ class Caches(_api_module.BaseModule):
   def list(
       self, *, config: Optional[types.ListCachedContentsConfigOrDict] = None
   ) -> Pager[types.CachedContent]:
+    """Lists cached contents.
+
+    Args:
+      config (ListCachedContentsConfig): Optional configuration for the list
+        request.
+
+    Returns:
+      A Pager object that contains one page of cached contents. When iterating
+      over
+      the pager, it automatically fetches the next page if there are more.
+
+    Usage:
+
+    .. code-block:: python
+      for cached_content in client.caches.list():
+        print(cached_content.name)
+    """
+
+    list_request = self._list
     return Pager(
         'cached_contents',
-        self._list,
+        list_request,
         self._list(config=config),
         config,
     )
@@ -1423,15 +1515,6 @@ class AsyncCaches(_api_module.BaseModule):
   async def _list(
       self, *, config: Optional[types.ListCachedContentsConfigOrDict] = None
   ) -> types.ListCachedContentsResponse:
-    """Lists cached content configurations.
-
-    .. code-block:: python
-
-      cached_contents = await client.aio.caches.list(config={'page_size': 2})
-      async for cached_content in cached_contents:
-        print(cached_content)
-    """
-
     parameter_model = types._ListCachedContentsParameters(
         config=config,
     )
@@ -1492,9 +1575,28 @@ class AsyncCaches(_api_module.BaseModule):
   async def list(
       self, *, config: Optional[types.ListCachedContentsConfigOrDict] = None
   ) -> AsyncPager[types.CachedContent]:
+    """Lists cached contents asynchronously.
+
+    Args:
+      config (ListCachedContentsConfig): Optional configuration for the list
+        request.
+
+    Returns:
+      A Pager object that contains one page of cached contents. When iterating
+      over
+      the pager, it automatically fetches the next page if there are more.
+
+    Usage:
+
+    .. code-block:: python
+      async for cached_content in await client.aio.caches.list():
+        print(cached_content.name)
+    """
+
+    list_request = self._list
     return AsyncPager(
         'cached_contents',
-        self._list,
+        list_request,
         await self._list(config=config),
         config,
     )