+++ google/genai/tests/types/test_part_type.py
@@ -52,6 +52,9 @@ def test_content_empty_parts_text():
 
 
 def test_two_candidates_text(caplog, generate_content_response):
+  from ... import types as types_module
+  types_module._response_text_warning_logged = False
+
   generate_content_response.candidates = [
       types.Candidate(
           content=types.Content(
@@ -545,3 +548,69 @@ def test_code_execution_result_one_candidate():
   )
 
   assert response.code_execution_result == '"hello"'
+
+
+def test_from_file_media_resolution_str():
+  file_uri = types.Part.from_uri(
+      file_uri='gs://test',
+      mime_type='image/png',
+      media_resolution='MEDIA_RESOLUTION_LOW',
+  )
+  assert file_uri.file_data.file_uri == 'gs://test'
+  assert file_uri.file_data.mime_type == 'image/png'
+  assert file_uri.media_resolution.level == types.PartMediaResolutionLevel.MEDIA_RESOLUTION_LOW
+
+
+def test_from_file_media_resolution_enum():
+  file_uri = types.Part.from_uri(
+      file_uri='gs://test',
+      mime_type='image/png',
+      media_resolution=types.PartMediaResolutionLevel.MEDIA_RESOLUTION_LOW,
+  )
+  assert file_uri.file_data.file_uri == 'gs://test'
+  assert file_uri.file_data.mime_type == 'image/png'
+  assert file_uri.media_resolution.level == types.PartMediaResolutionLevel.MEDIA_RESOLUTION_LOW
+
+
+def test_from_file_media_resolution_object():
+  file_uri = types.Part.from_uri(
+      file_uri='gs://test',
+      mime_type='image/png',
+      media_resolution=types.PartMediaResolution(level='MEDIA_RESOLUTION_LOW'),
+  )
+  assert file_uri.file_data.file_uri == 'gs://test'
+  assert file_uri.file_data.mime_type == 'image/png'
+  assert file_uri.media_resolution.level == types.PartMediaResolutionLevel.MEDIA_RESOLUTION_LOW
+
+
+def test_from_bytes_media_resolution_str():
+  file_uri = types.Part.from_bytes(
+      data=b'1234',
+      mime_type='image/png',
+      media_resolution='MEDIA_RESOLUTION_LOW',
+  )
+  assert file_uri.inline_data.data == b'1234'
+  assert file_uri.inline_data.mime_type == 'image/png'
+  assert file_uri.media_resolution.level == types.PartMediaResolutionLevel.MEDIA_RESOLUTION_LOW
+
+
+def test_from_bytes_media_resolution_enum():
+  file_uri = types.Part.from_bytes(
+      data=b'1234',
+      mime_type='image/png',
+      media_resolution=types.PartMediaResolutionLevel.MEDIA_RESOLUTION_LOW,
+  )
+  assert file_uri.inline_data.data == b'1234'
+  assert file_uri.inline_data.mime_type == 'image/png'
+  assert file_uri.media_resolution.level == types.PartMediaResolutionLevel.MEDIA_RESOLUTION_LOW
+
+
+def test_from_bytes_media_resolution_object():
+  file_uri = types.Part.from_bytes(
+      data=b'1234',
+      mime_type='image/png',
+      media_resolution=types.PartMediaResolution(level='MEDIA_RESOLUTION_LOW'),
+  )
+  assert file_uri.inline_data.data == b'1234'
+  assert file_uri.inline_data.mime_type == 'image/png'
+  assert file_uri.media_resolution.level == types.PartMediaResolutionLevel.MEDIA_RESOLUTION_LOW